Fixup Marketplace builds for Win64/Mac/iOS/Android
- Add NDK as native module
   Allows for NDK modification without needing to comply with restrictions with plugin source files.
   Allows for including and using NDK code without needing to macro export. Simply include THIRD_PARTY_INCLUDES_START/END instead of adding export macro in base NDK.

- Client & Editor Client
   Depend on NDK, should be only modules that directly depend on NDK.
   Use non-Unreal types in Private implementation. Use Unreal types in Public implementation. (Allows other modules to depend on these without gaining reference to non-Unreal types).
   Act as translation between Unreal/NDK.
- Implementation
   Depend on Client, but not NDK, sample implementation of Inworld.
   Use only Unreal types.

+++ InworldAI/Source/InworldAIIntegration/Public/InworldPlayerAudioCaptureComponent.h
@@ -8,7 +8,6 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "InworldUtils.h"
 #include "AudioCaptureCore.h"
 #include "AudioDevice.h"
 #include "Containers/ContainerAllocationPolicies.h"
@@ -27,9 +26,9 @@ struct FPlayerVoiceCaptureInfoRep
     GENERATED_BODY()
 
     UPROPERTY()
-	TArray<int16> MicSoundData;
+	TArray<uint8> MicSoundData;
     UPROPERTY()
-	TArray<int16> OutputSoundData;
+	TArray<uint8> OutputSoundData;
 };
 
 UCLASS(ClassGroup = (Inworld), meta = (BlueprintSpawnableComponent))
